FIx some bugs
Some bugs are fixed this time. Meanwhile, I use double instead of float
as much as possible to increase the accuracy. I also choose to compute
Q each time I want to compute the cost of an edge instead of storing
the QList for each vertex. This also can somehow increase the accuracy.